+++ CHANGELOG.md
@@ -1,6 +1,9 @@
+## 1.0.7
+- Add toPem and fromPem to RSAPublicKey
+- Add toPem and fromPem to RSAPrivateKey
+
 ## 1.0.6
 - Update dependencies
-
 
 ## 1.0.5
 - Add more Documentation and fix a small bug